Fix PHP 8 warning in TagManagerListener.php
When you use the tags bundle with PHP 8 in debug mode, you get a warning, that $GLOBALS['TL_JAVASCRIPT'] was not found. In PHP 7 it was just a notice that was ignored in debug mode. The pull request contains a check for the $GLOBALS['TL_JAVASCRIPT'] and shortens the check for jquery.